ATCToeDecode work and converted the ASCHipForce controller to the new system.

+++ software/atrias_controllers/asc_toe_decode/include/asc_toe_decode/ASCToeDecode.hpp
@@ -39,7 +39,7 @@ class ASCToeDecode : public AtriasController {
 		  *
 		  * Call this function every millisecond with new data.
 		  */
-		double operator()(double force);
+		double operator()(uint16_t force);
 
 		/**
 		  * @brief This returns true if the toe is contacting the ground.
